+++ testcontainers/src/main/java/org/jdbi/v3/testing/junit5/tc/package-info.java
@@ -1,0 +1,18 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Jdbi test support for <a href="https://testcontainers.org/">Testcontainer based JDBC containers</a>. Any
+ * database that is supported can be used as a test database in <a href="https://junit.org/junit5/">JUnit 5</a> tests.
+ */
+package org.jdbi.v3.testing.junit5.tc;